+++ crates/miden-tx/src/testing/mock_chain/mod.rs
@@ -28,7 +28,8 @@ use miden_objects::{
     testing::account_code::DEFAULT_AUTH_SCRIPT,
     transaction::{
         ChainMmr, ExecutedTransaction, InputNote, InputNotes, OutputNote, ProvenTransaction,
-        ToInputNoteCommitments, TransactionId, TransactionInputs, TransactionScript,
+        ToInputNoteCommitments, TransactionHeader, TransactionId, TransactionInputs,
+        TransactionScript,
     },
 };
 use rand::{Rng, SeedableRng};
@@ -463,7 +464,7 @@ impl MockChain {
     /// This method does not modify the chain state.
     pub fn prove_transaction_batch(&self, proposed_batch: ProposedBatch) -> ProvenBatch {
         let (
-            _transactions,
+            transactions,
             block_header,
             _chain_mmr,
             _unauthenticated_note_proofs,
@@ -474,6 +475,20 @@ impl MockChain {
             batch_expiration_block_num,
         ) = proposed_batch.into_parts();
 
+        let verified_txs = transactions
+            .iter()
+            .map(|proven_tx| {
+                TransactionHeader::new_unchecked(
+                    proven_tx.id(),
+                    proven_tx.account_id(),
+                    proven_tx.account_update().initial_state_commitment(),
+                    proven_tx.account_update().final_state_commitment(),
+                    proven_tx.input_notes().iter().map(|note| note.nullifier()).collect(),
+                    proven_tx.output_notes().iter().map(|note| note.id()).collect(),
+                )
+            })
+            .collect();
+
         ProvenBatch::new_unchecked(
             id,
             block_header.commitment(),
@@ -482,6 +497,7 @@ impl MockChain {
             input_notes,
             output_notes,
             batch_expiration_block_num,
+            verified_txs,
         )
     }
 
@@ -897,6 +913,9 @@ impl MockChain {
                 self.pending_objects.updated_accounts.clone(),
                 self.pending_objects.output_note_batches.clone(),
                 self.pending_objects.created_nullifiers.clone(),
+                // TODO: For now we can't easily compute the verified transactions of this block.
+                // Let's do this as part of miden-base/#1224.
+                vec![],
             );
 
             for (batch_index, note_batch) in